Add has_error property and tests
