+++ pypaperless/models/mixins/models/__init__.py
@@ -1,7 +1,7 @@
 """Mixins for PyPaperless models."""
 
 from .creatable import CreatableMixin
-from .data_fields import MatchingFieldsMixin  # , PermissionFieldsMixin
+from .data_fields import MatchingFieldsMixin
 from .deletable import DeletableMixin
 from .securable import SecurableDraftMixin, SecurableMixin
 from .updatable import UpdatableMixin
@@ -10,7 +10,6 @@ __all__ = (
     "CreatableMixin",
     "DeletableMixin",
     "MatchingFieldsMixin",
-    # "PermissionFieldsMixin",
     "UpdatableMixin",
     "SecurableMixin",
     "SecurableDraftMixin",
